Allow a thing to be used for DragonConnect and DragonPulse

+++ admin/lib/things/mgmt.js
@@ -5,12 +5,11 @@ var fs = require('fs'),
 
 var AWS = require('aws-sdk'),
     Bluebird = require('bluebird'),
+    mkdirp = require('mkdirp'),
     rimraf = require('rimraf');
 
 var config = require('dragonpulse-config'),
     configureAws = require('./../util/helper').configureAws;
-
-var SimpleError = require('./../errors/simpleerror');
 
 var REGISTRY_BASE = './../../registry';
 var KEYS_AND_CERTIFICATES_FILE_NAME = 'certificatesAndKeys.json';
@@ -19,48 +18,97 @@ function getRegistry(thingId) {
   return path.join(path.relative(process.cwd(), path.dirname(module.filename)), REGISTRY_BASE, thingId);
 }
 
+function getThing(iot, context) {
+  var awsDescribeThing = Bluebird.promisify(iot.describeThing, { context: iot });
+
+  var params = {
+    thingName: context.thingId
+  };
+
+  return awsDescribeThing(params)
+    .then(function(thing) {
+        return {
+          thingId: thing.thingName,
+          attributes: thing.attributes
+        }
+      });
+}
+
+function getOrCreateThing(iot, context) {
+  return getThing(iot, context)
+    .catch(function(err) {
+      switch(err.statusCode) {
+        case 404: {
+          // The thing does not exist
+          return createThing(iot, context);
+        }
+        default: {
+          throw err;
+        }
+      }
+    })
+}
+
 function createThing(iot, context) {
   console.info('Creating thing ' + context.thingId);
 
   var awsCreateThing = Bluebird.promisify(iot.createThing, { context: iot });
 
   var params = {
+    thingName: context.thingId
+  };
+
+  return awsCreateThing(params)
+    .then(function() {
+        return createPrincipal(iot, context);
+      })
+    .then(function(keyAndCertificates) {
+        return attachPrincipalToThing(iot, keyAndCertificates.certificateArn, context)
+          .then(function() {
+              return updateThingAttributes(iot, keyAndCertificates, context);
+            })
+          .catch(function(err) {
+              throw err;
+            });
+      });
+}
+
+function updateThingAttributes(iot, keyAndCertificate, context) {
+  var awsUpdateThing = Bluebird.promisify(iot.updateThing, { context: iot });
+
+  var attributes = {
+    reference: 'ArrowDragonConnectAndDragonPulse',
+    certificateArn: keyAndCertificate.certificateArn
+  };
+
+  var params = {
     thingName: context.thingId,
     attributePayload: {
-      attributes: {
-        reference: 'DragonPulse'
-      }
+      attributes: attributes
     }
   };
 
-  return Bluebird.resolve()
+  return awsUpdateThing(params)
     .then(function() {
-        return awsCreateThing(params)
+        return {
+          thingId: context.thingId,
+          attributes: attributes
+        }
       })
     .catch(function(err) {
-        switch(err.statusCode) {
-          case 409: {
-            throw new SimpleError('Resource Already Exists', err);
-          }
-          default: {
-            throw new SimpleError('Unknown Condition', err)
-          }
-        }
+        throw err;
       });
 }
 
 function createPrincipal(iot, context) {
   console.info('Creating principal');
 
-  var mkdir = Bluebird.promisify(fs.mkdir);
+  var mkdir = Bluebird.promisify(mkdirp);
   var awsCreateKeysAndCertificate = Bluebird.promisify(iot.createKeysAndCertificate, { context: iot });
 
   var thingRegistry = getRegistry(context.thingId);
 
-  return Bluebird.resolve()
-    .then(function() {
-        mkdir(thingRegistry);
-      })
+  return mkdir(thingRegistry)
     .then(function() {
         return awsCreateKeysAndCertificate({
             setAsActive: true
@@ -90,19 +138,16 @@ function attachPrincipalToThing(iot, certificateArn, context) {
     })
 }
 
-function attachPolicyToPrincipal(iot, certificateArn, context) {
+function attachPolicyToPrincipal(iot, thing) {
   var policyName = config.iot.policies.DragonPulseThing;
 
-  console.info('Attaching policy ' + policyName + ' to principal for thing ' + context.thingId);
+  console.info('Attaching policy ' + policyName + ' to principal for thing ' + thing.thingId);
 
   var awsAttachPrincipalPolicy = Bluebird.promisify(iot.attachPrincipalPolicy, { context: iot });
 
-  return Bluebird.resolve()
-    .then(function() {
-        return awsAttachPrincipalPolicy({
-          policyName: policyName,
-          principal: certificateArn
-        })
+  return awsAttachPrincipalPolicy({
+        policyName: policyName,
+        principal: thing.attributes.certificateArn
       });
 }
 
@@ -111,22 +156,13 @@ function createResources(context) {
 
   var iot = new AWS.Iot();
 
-  return Bluebird.resolve()
-    .then(function () {
-        return createThing(iot, context)
-      })
-    .then(function () {
-        return createPrincipal(iot, context)
-      })
-    .then(function(data) {
-        return attachPrincipalToThing(iot, data.certificateArn, context)
-          .then(function() {
-            return attachPolicyToPrincipal(iot, data.certificateArn, context);
-          })
+  return getOrCreateThing(iot, context)
+    .then(function(thing) {
+        return attachPolicyToPrincipal(iot, thing);
       })
     .catch(function(err) {
         console.error('ERROR');
-        console.error('  Condition:  ' + err.condition + ', Message:  ' + err.message);
+        console.error('  Message:  ' + err.message);
       });
 }
 
@@ -136,12 +172,9 @@ function deletePrincipal(iot, certificateId, context) {
   var awsUpdateCertificate = Bluebird.promisify(iot.updateCertificate, { context: iot });
   var awsDeleteCertificate = Bluebird.promisify(iot.deleteCertificate, { context: iot });
 
-  return Bluebird.resolve()
-    .then(function() {
-        return awsUpdateCertificate({
-          certificateId: certificateId,
-          newStatus: 'INACTIVE'
-        })
+  return awsUpdateCertificate({
+        certificateId: certificateId,
+        newStatus: 'INACTIVE'
       })
     .then(function() {
         return awsDeleteCertificate({
@@ -155,11 +188,8 @@ function deleteThing(iot, context) {
 
   var awsDeleteThing = Bluebird.promisify(iot.deleteThing, { context: iot });
 
-  return Bluebird.resolve()
-    .then(function() {
-        return awsDeleteThing({
-          thingName: context.thingId
-        })
+  return awsDeleteThing({
+        thingName: context.thingId
       });
 }
 
@@ -168,29 +198,35 @@ function detachPrincipalFromThing(iot, certificateArn, context) {
 
   var awsDetachThingPrincipal = Bluebird.promisify(iot.detachThingPrincipal, { context: iot });
 
-  return Bluebird.resolve()
-    .then(function() {
-        return awsDetachThingPrincipal({
-          principal: certificateArn,
-          thingName: context.thingId
-        })
+  return awsDetachThingPrincipal({
+        principal: certificateArn,
+        thingName: context.thingId
       });
 }
 
-function detachPolicyFromPrincipal(iot, certificateArn, context) {
-  var policyName = config.iot.policies.DragonPulseThing;
+function detachPoliciesFromPrincipal(iot, certificateArn, context) {
 
-  console.info('Detaching policy ' + policyName + ' from principal for thing ' + context.thingId);
-
+  var awsListPrincipalPolicies = Bluebird.promisify(iot.listPrincipalPolicies, { context: iot });
   var awsDetachPrincipalPolicy = Bluebird.promisify(iot.detachPrincipalPolicy, { context: iot });
 
-  return Bluebird.resolve()
-    .then(function() {
-        return awsDetachPrincipalPolicy({
-          policyName: policyName,
-          principal: certificateArn
+  return awsListPrincipalPolicies({
+        principal: certificateArn
+      })
+    .then(function(result) {
+      Bluebird.each(result.policies, function(policyRef) {
+          var policyName = policyRef.policyName;
+
+          console.info('Detaching policy ' + policyName + ' from principal for thing ' + context.thingId);
+
+          return awsDetachPrincipalPolicy({
+            policyName: policyName,
+            principal: certificateArn
+          })
         })
-      });
+      .catch(function (err) {
+        throw err;
+      })
+    });
 }
 
 function deleteResources(context) {
@@ -201,23 +237,37 @@ function deleteResources(context) {
   var iot = new AWS.Iot();
 
   var thingRegistry = getRegistry(context.thingId);
-  var keysAndCertificates = require(path.join(REGISTRY_BASE, context.thingId, KEYS_AND_CERTIFICATES_FILE_NAME));
 
-  return Bluebird.resolve()
-    .then(function() {
-        return detachPolicyFromPrincipal(iot, keysAndCertificates.certificateArn, context);
+  return getThing(iot, context)
+    .then(function(thing) {
+        if (!thing.attributes.hasOwnProperty('certificateArn')) {
+          throw new ReferenceError('Thing ' + context.thingId + ' was not created properly.  ' +
+            'It should have an attribute of certificateArn.');
+        }
+
+        var certificateArn = thing.attributes.certificateArn;
+        var certificateId = /.*cert\/(.*)/.exec(certificateArn)[1];
+
+        detachPoliciesFromPrincipal(iot, certificateArn, context)
+          .then(function() {
+              return detachPrincipalFromThing(iot, certificateArn, context);
+            })
+          .then(function() {
+              return deletePrincipal(iot, certificateId, context);
+            })
+          .then(function() {
+              return deleteThing(iot, context)
+            })
+          .then(function() {
+              rmdir(thingRegistry);
+            })
+          .catch(function(err) {
+              throw err;
+            });
       })
-    .then(function() {
-        return detachPrincipalFromThing(iot, keysAndCertificates.certificateArn, context);
-      })
-    .then(function() {
-        return deletePrincipal(iot, keysAndCertificates.certificateId, context);
-      })
-    .then(function() {
-        return deleteThing(iot, context)
-      })
-    .then(function() {
-        rmdir(thingRegistry);
+    .catch(function(err) {
+        console.error('ERROR');
+        console.error('  Message:  ' + err.message);
       });
 }
